+++ flash_examples/graph_classification.py
@@ -14,13 +14,12 @@
 import torch
 
 import flash
-from flash.core.utilities.imports import _TORCH_GEOMETRIC_AVAILABLE
+from flash.core.utilities.imports import example_requires
 from flash.graph import GraphClassificationData, GraphClassifier
 
-if _TORCH_GEOMETRIC_AVAILABLE:
-    from torch_geometric.datasets import TUDataset
-else:
-    raise ModuleNotFoundError("Please, pip install -e '.[graph]'")
+example_requires("graph")
+
+from torch_geometric.datasets import TUDataset  # noqa: E402
 
 # 1. Create the DataModule
 dataset = TUDataset(root="data", name="KKI")